scraper/requirements.txt
telethon
python-dotenv
asyncio
psycopg2-binary
requests
